:ok_hand: IMPROVE: Use Sphinx HTML assets policy to decide whether include assets
Sphinx v4.1.0 will include `Sphinx.registry.html_assets_policy` that defines the
policy to whether include or not HTML assets in pages where the extension is not
used.

This PR makes usage of this policy to decide if sphinx-tabs assets should be
included or not in the page that's being rendered.

Reference: https://github.com/sphinx-doc/sphinx/issues/9115