[fix] prefers-color-scheme: use dark theme if body[data-theme="auto"]
Before this patch the CSS implementation assumes that "dark" is supported by the
theme if the `data-theme` attribute of the <body> tag is unset.

Since most themes are using "light" colors (compare https://sphinx-themes.org)
and do not set `data-theme` it is better to assume light is the default, even
when the browser setting `prefers-color-scheme: dark`!

BTW: remove duplication of styles from dark/light theme that are already set in
the common style.

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>